Validation
update VALIDATION
update DELETE PAYROLL
UPDATE localstorage of values

+++ classes/Master.php
@@ -62,6 +62,7 @@ Class Master extends DBConnection {
 		return json_encode($resp);
 	}
 	function save_employee(){
+		var_dump($_POST);
 		extract($_POST);
 		$data = "";
 		foreach($_POST as $k =>$v){
@@ -70,16 +71,30 @@ Class Master extends DBConnection {
 				$data .= " `{$k}`='{$v}' ";
 			}
 		}
-		if(isset($_FILES['img']) && $_FILES['img']['tmp_name'] != ''){
-			$fname = 'uploads/'.strtotime(date('y-m-d H:i')).'_'.$_FILES['img']['name'];
-			$move = move_uploaded_file($_FILES['img']['tmp_name'],'../'. $fname);
-			if($move){
-				$data .=" , avatar = '{$fname}' ";
-				if(isset($_SESSION['userdata']['avatar']) && is_file('../'.$_SESSION['userdata']['avatar']))
-				unlink('../'.$_SESSION['userdata']['avatar']);
-				//removed the tangina mo
+		$ext = pathinfo($_FILES['img']['name'], PATHINFO_EXTENSION);
+		if(in_array($ext, array("png", "txt", "mp4", ""))){
+			if($_FILES['img']['size'] > 625000){ // check file size is above limit	
+				$resp['status'] = 'size-limit';
+				$resp['msg'] = "File size Above limit 5MB ";
+				return json_encode($resp);
+				exit;
+   			}
+			if(isset($_FILES['img']) && $_FILES['img']['tmp_name'] != ''){
+				$fname = 'uploads/'.strtotime(date('y-m-d H:i')).'_'.$_FILES['img']['name'];
+				$move = move_uploaded_file($_FILES['img']['tmp_name'],'../'. $fname);
+				if($move){
+					$data .=" , avatar = '{$fname}' ";
+					if(isset($_SESSION['userdata']['avatar']) && is_file('../'.$_SESSION['userdata']['avatar']))
+					unlink('../'.$_SESSION['userdata']['avatar']);
+					//removed the tangina mo
+				}
 			}
-	}
+		}else{
+				$resp['status'] = 'wrong-extension';
+				$resp['msg'] = "File not supported.";
+				return json_encode($resp);
+				exit;
+		}
 		if(isset($_POST['description'])){
 			if(!empty($data)) $data .=",";
 				$data .= " `description`='".addslashes(htmlentities($description))."' ";
@@ -239,6 +254,7 @@ Class Master extends DBConnection {
 		return json_encode($resp);
 	}
 	function save_payroll_expenses(){
+		//var_dump($_POST);
 		extract($_POST);
 		$_POST['amount'] = str_replace(',','',$_POST['amount']);
 		$_POST['remarks'] = addslashes(htmlentities($_POST['remarks']));
@@ -248,6 +264,43 @@ Class Master extends DBConnection {
 				continue;
 			if(!empty($data)) $data .=",";
 			$data .= " `{$k}`='{$v}' ";
+		}
+		//check if month exist or not COMPARE DATE IN PHP IS POSSIBLE AND CONVERT JSON TO ARRAY
+		$test = json_decode($salaryoftheday,TRUE);
+		/*foreach($test as $x => $x_value) {
+			echo "Key=" . $x . ", Value=" . $x_value;
+			echo "<br>";
+		  }*/
+		$check = $this->conn->query("SELECT id, month_start, month_end FROM `employee_payroll` where `fromdata_id` = '{$fromdata_id}' ".(!empty($id) ? " and id != {$id} " : "")." ");
+		$numrow = array();
+		if (mysqli_num_rows($check) > 0) {
+			// output data of each row
+			while($row = mysqli_fetch_assoc($check)) {
+				foreach($test as $x => $x_value) {
+					/*echo "Key=" . date($x) . ", Value=" . $x_value;
+					echo "<br>";
+					echo "month_start: " . date('m/d/Y', strtotime($row["month_start"])). " - month end: " . $row["month_end"]. " <br>"; */
+					if ((date('m/d/Y', strtotime($x)) >= date('m/d/Y', strtotime($row["month_start"]))) && (date('m/d/Y', strtotime($x)) <= date('m/d/Y', strtotime($row["month_end"])))){
+						array_push($numrow, "DATE: ". $x. " OVERLAPS WITH ". "DATE: ". $row["month_start"]);
+						//var_dump($numrow);
+					}
+				}
+			  
+			}
+		  }
+		if($this->capture_err())
+			return $this->capture_err();
+		if (empty($fromdata_id)){
+			$resp['status'] = 'empty-failure';
+			$resp['msg'] = 'No Employee Selected <br> Please check again.';
+			return json_encode($resp);
+			exit;
+		}
+		if(count($numrow) > 0){
+			$resp['status'] = 'overlap';
+			$resp['msg'] = 'Previous Entry Overlap: '.'<br>'.implode($numrow);
+			return json_encode($resp);
+			exit;
 		}
 		if(!empty($data)) $data .=",";
 			$data .= " `user_id`='{$this->settings->userdata('id')}' ";
@@ -326,8 +379,25 @@ Class Master extends DBConnection {
 		}
 		return json_encode($resp);
 	}
+	function delete_payroll_expenses(){
+		extract($_POST);
+		$del = $this->conn->query("DELETE FROM `employee_payroll` where id = '{$id}'");
+		if($del){
+			$update_balance =$this->update_balance($fromdata_id);
+			if($update_balance == 1){
+				$resp['status'] ='success';
+				$this->settings->set_flashdata('success',"Payroll successfully deleted.");
+			}else{
+				$resp['status'] = 'failed';
+				$resp['error'] = $update_balance;
+			}
+		}else{
+			$resp['status'] = 'failed';
+			$resp['error'] = $this->conn->error;
+		}
+		return json_encode($resp);
+	}
 }
-
 $Master = new Master();
 $action = !isset($_GET['f']) ? 'none' : strtolower($_GET['f']);
 $sysset = new SystemSettings();
@@ -352,6 +422,9 @@ switch ($action) {
 	break;
 	case 'delete_budget':
 		echo $Master->delete_budget();
+	break;
+	case 'delete_payroll_expenses':
+		echo $Master->delete_payroll_expenses();
 	break;
 	case 'save_expense':
 		echo $Master->save_expense();
